# Conflicts:
#	ImageSource.podspec

+++ ImageSource/Remote/CachingImageDownloader.swift
@@ -1,7 +1,7 @@
 public protocol ImageDownloader {
     func downloadImageAtUrl(
         _: URL,
-        progressHandler: ((_ receivedSize: Int, _ expectedSize: Int) -> ())?,
+        progressHandler: ((_ receivedSize: Int64, _ expectedSize: Int64) -> ())?,
         completion: @escaping (_ image: CGImage?, _ error: Error?) -> ()
     ) -> CancellableImageDownload
 }
